Move pkg/jsonmessage to client/pkg/jsonmessage
