+++ integration/internal/image/load.go
@@ -10,9 +10,9 @@ import (
 	"testing"
 
 	"github.com/docker/docker/internal/testutils/specialimage"
-	"github.com/docker/docker/pkg/jsonmessage"
 	"github.com/moby/go-archive"
 	"github.com/moby/moby/client"
+	"github.com/moby/moby/client/pkg/jsonmessage"
 	"gotest.tools/v3/assert"
 )
